Added edge weights to directed graphs
